requirements.txt
lxml==5.3.0
python-docx==1.1.2
python-docx-replace==0.4.4
typing_extensions==4.12.2
